requirements.txt
discord.py
gspread
google-auth
python-dotenv 